+++ confik-macros/tests/trybuild/27-field-access.rs
@@ -1,0 +1,16 @@
+//! Check that we can reference builder fields
+use confik::Configuration;
+
+#[derive(Configuration, Debug, PartialEq)]
+struct Config {
+    #[confik(default)]
+    param: String,
+}
+
+type Builder = <Config as Configuration>::Builder;
+
+fn main() {
+    let _ = Builder {
+        param: Default::default(),
+    };
+}