+++ confik/src/sources/json_source.rs
@@ -25,38 +25,20 @@ impl<'a> JsonSource<'a> {
     }
 }
 
-impl<'a> Source for JsonSource<'a> {
+impl<T: ConfigurationBuilder> Source<T> for JsonSource<'_> {
     fn allows_secrets(&self) -> bool {
         self.allow_secrets
     }
 
-    fn provide<T: ConfigurationBuilder>(&self) -> Result<T, Box<dyn Error + Sync + Send>> {
+    fn provide(&self) -> Result<T, Box<dyn Error + Sync + Send>> {
         Ok(serde_json::from_str(&self.contents)?)
     }
 }
 
-impl<'a> fmt::Debug for JsonSource<'a> {
+impl fmt::Debug for JsonSource<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("JsonSource")
             .field("allow_secrets", &self.allow_secrets)
             .finish_non_exhaustive()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn defaults() {
-        let source = JsonSource::new("{}");
-        assert!(!source.allows_secrets());
-    }
-
-    #[test]
-    fn clone() {
-        let source = JsonSource::new("{}").allow_secrets();
-        assert!(source.allows_secrets());
-        assert!(source.clone().allow_secrets);
     }
 }