+++ confik/tests/unkeyed_containers/mod.rs
@@ -2,13 +2,15 @@ macro_rules! create_tests_for {
     ($container:ty) => {
         use confik::Configuration;
 
+        #[allow(dead_code)] // unused in no-default-features cases
         #[derive(Debug, Configuration, PartialEq, Eq, Hash, Ord, PartialOrd)]
-        #[confik(derive(Hash, PartialEq, Eq, Ord, PartialOrd))]
+        #[confik(forward(derive(Hash, PartialEq, Eq, Ord, PartialOrd)))]
         struct TwoVals {
             first: usize,
             second: usize,
         }
 
+        #[allow(dead_code)] // unused in no-default-features cases
         #[derive(Debug, Configuration, PartialEq, Eq)]
         struct Target {
             val: $container,